refactor:firebase関連のlibのフォルダ階層を整理
