+++ src/lib/firebase/config.ts
